backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
aiomysql==0.2.0
cryptography==42.0.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
scikit-learn==1.5.0
scipy==1.13.1
numpy==1.26.4
pandas==2.2.2
apscheduler==3.10.4
twilio==9.2.0
pydantic[email]==2.7.1
pydantic-settings==2.3.1
python-multipart==0.0.9
websockets==12.0
boto3==1.34.0
statsmodels==0.14.2         # STL decomposition + Holt-Winters forecast (Upgrade 2)
# torch==2.3.0              # LSTM Autoencoder (Phase 2 — install separately: pip install torch)
